+++ src/org/scauhci/ExamSystem/android/dao/QuestionOptionDao.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import org.scauhci.ExamSystem.android.pojo.CoursePojo;
 import org.scauhci.ExamSystem.android.pojo.QuestionOptionPojo;
 import org.scauhci.ExamSystem.android.pojo.QuestionPojo;
-import org.scauhci.ExamSystem.android.tool.ExecuteResultFlag;
+import org.scauhci.ExamSystem.android.tool.Flag;
 import org.scauhci.ExamSystem.android.tool.HashValue;
 
 import android.database.Cursor;
@@ -18,9 +18,7 @@ public class QuestionOptionDao {
 	String tableName = "question_option";
 	private QuestionOptionPojo latestQuestionOptionPojo = null;
 
-	public int add(QuestionOptionPojo questionOptionPojo) {
-		int executeResult = ExecuteResultFlag.ERROR;
-
+	public QuestionOptionPojo add(QuestionOptionPojo questionOptionPojo) {
 		long questionOptionId;
 		for (questionOptionId = HashValue.getDJBHashValue(questionOptionPojo
 				.getQuestionOptionContent()); getQuestionOptionPojoByQuestionOptionId(Long
@@ -38,23 +36,22 @@ public class QuestionOptionDao {
 
 		daoHelper.insert(tableName, keys, values);
 
-		return executeResult;
+		return questionOptionPojo;
 	}
 
-	public int delete(QuestionOptionPojo questionOptionPojo) {
-		int executeResult = ExecuteResultFlag.ERROR;
+	public QuestionOptionPojo delete(QuestionOptionPojo questionOptionPojo) {
 
-		String[] keys = { "questionOptionId" };
-		String[] values = { questionOptionPojo.getQuestionOptionId() };
+		if ((questionOptionPojo = completeQuestionOptionPojo(questionOptionPojo)) != null) {
+			String[] keys = { "questionOptionId" };
+			String[] values = { questionOptionPojo.getQuestionOptionId() };
 
-		daoHelper.delete(tableName, keys, values);
+			daoHelper.delete(tableName, keys, values);
+		}
 
-		return executeResult;
+		return questionOptionPojo;
 	}
 
-	public int change(QuestionOptionPojo questionOptionPojo) {
-		int executeResult = ExecuteResultFlag.ERROR;
-
+	public QuestionOptionPojo change(QuestionOptionPojo questionOptionPojo) {
 		HashMap<String, String> keyValueMap = getKeyValueMapByQuestionOptionPojo(questionOptionPojo);
 
 		String[] keys = new String[keyValueMap.size()];
@@ -67,13 +64,19 @@ public class QuestionOptionDao {
 
 		daoHelper.update(tableName, keys, newValues, whereConditionKeys,
 				whereConditionValues);
+		questionOptionPojo = completeQuestionOptionPojo(questionOptionPojo);
 
-		return executeResult;
+		return questionOptionPojo;
 	}
 
-	public HashMap<String, String> getKeyValueMapByQuestionOptionPojo(QuestionOptionPojo questionOptionPojo){
+	public HashMap<String, String> getKeyValueMapByQuestionOptionPojo(
+			QuestionOptionPojo questionOptionPojo) {
 		HashMap<String, String> keyValueMap = new HashMap<String, String>();
 
+		if (questionOptionPojo.getQuestionOptionId() != null) {
+			keyValueMap.put("questionOptionId",
+					questionOptionPojo.getQuestionOptionId());
+		}
 		if (questionOptionPojo.getQuestionId() != null) {
 			keyValueMap.put("questionId", questionOptionPojo.getQuestionId());
 		}
@@ -83,10 +86,10 @@ public class QuestionOptionDao {
 		}
 		keyValueMap.put("isQuestionStdAnswer",
 				questionOptionPojo.isQuestionStdAnswer() + "");
-		
+
 		return keyValueMap;
 	}
-	
+
 	public QuestionOptionPojo getQuestionOptionPojoByQuestionOptionId(
 			String questionOptionId) {
 		QuestionOptionPojo questionOptionPojo = new QuestionOptionPojo();
@@ -97,7 +100,7 @@ public class QuestionOptionDao {
 
 		Cursor questionOptionCursor = daoHelper.select(tableName, keys,
 				whereConditionKeys, whereConditionValues);
-		while (questionOptionCursor.moveToNext()) {
+		if (questionOptionCursor.moveToFirst()) {
 			questionOptionPojo.setQuestionOptionId(questionOptionCursor
 					.getString(questionOptionCursor
 							.getColumnIndex("questionOptionId")));
@@ -108,12 +111,15 @@ public class QuestionOptionDao {
 					.getBoolean(questionOptionCursor
 							.getString(questionOptionCursor
 									.getColumnIndex("questionOptionContent"))));
+		} else {
+			questionOptionPojo = null;
 		}
 
 		return questionOptionPojo;
 	}
-	
-	public QuestionOptionPojo completeQuestionOptionPojo(QuestionOptionPojo questionOptionPojo) {
+
+	public QuestionOptionPojo completeQuestionOptionPojo(
+			QuestionOptionPojo questionOptionPojo) {
 		HashMap<String, String> keyValueMap = getKeyValueMapByQuestionOptionPojo(questionOptionPojo);
 
 		String[] keys = { "*" };
@@ -124,7 +130,7 @@ public class QuestionOptionDao {
 
 		Cursor questionOptionCursor = daoHelper.select(tableName, keys,
 				whereConditionKeys, whereConditionValues);
-		while (questionOptionCursor.moveToNext()) {
+		if (questionOptionCursor.moveToFirst()) {
 			questionOptionPojo.setQuestionOptionId(questionOptionCursor
 					.getString(questionOptionCursor
 							.getColumnIndex("questionOptionId")));
@@ -135,8 +141,10 @@ public class QuestionOptionDao {
 					.getBoolean(questionOptionCursor
 							.getString(questionOptionCursor
 									.getColumnIndex("questionOptionContent"))));
+		} else {
+			questionOptionPojo = null;
 		}
-		
+
 		return questionOptionPojo;
 	}
